Moved running the pipeline into another activity
Unable to access the mainActBinding, but the code is cleaner. Finishing
the other activity to return to the Main UI is necessary.
