+++ mobile_app/app/src/main/AndroidManifest.xml
@@ -21,6 +21,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".RunPipelineActivity">
+        </activity>
     </application>
 
 </manifest>
